IMPALA-6820: Remove _impala_builtins from catalogd
The _impala_builtins database is initialized in the constructor of
Catalog and hence is inherited by both the CatalogServiceCatalog and
ImpaladCatalog. Since  _impala_builtins is not used by the catalog
server and to avoid the overhead of managing this database during normal
metadata operations (e.g. invalidate), it is moved to the ImpaladCatalog
class.

Change-Id: I166d8086db1d2920408f38dc56fe7c70a4c143a8
Reviewed-on: http://gerrit.cloudera.org:8080/9947
Reviewed-by: Tianyi Wang <twang@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 17137d783c24d0ec0827db62bb256ec4c2c05164)
Reviewed-on: http://gerrit.cloudera.org:8080/10048
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>